% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ejemplo-Colas.R
\name{colas_function}
\alias{colas_function}
\title{Ejemplo Proyecto Teoria de Colas}
\usage{
colas_function(lambda, mu, n)
}
\arguments{
\item{colas}{Resolver ejercicios como ser M/M/1 o M/M/k.}
}
\description{
Este paquete permite al usuario resolver ciertos ejercicios de Teoria de Colas mediante la Simulacion.
}
\examples{
colas_function()
}
\keyword{colas}
